requirements.txt
prefect
requests
pandas
awswrangler
python-dotenv
json